Add checks for disks and config to prevent a locked Linode from being in an undesired state (shutdown) during deletion attempts

+++ linode/instancedisk/framework_resource.go
@@ -323,6 +323,21 @@ func (r *Resource) Delete(
 		return
 	}
 
+	locked, err := helper.LinodeIsLockedWithCannotDeleteSubresources(ctx, client, linodeID)
+	if err != nil {
+		resp.Diagnostics.AddError("Failed to Get Locks of Linode", err.Error())
+		return
+	}
+
+	if locked {
+		resp.Diagnostics.AddError(
+			fmt.Sprintf("Can't Delete Disk %d in Linode %d", id, linodeID),
+			"The resource lock on the Linode prohibits deletion "+
+				"of its subresources, which includes this disk.",
+		)
+		return
+	}
+
 	deleteTimeout, diags := state.Timeouts.Delete(ctx, DefaultVolumeDeleteTimeout)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
